% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{a_line}
\alias{a_line}
\title{Line}
\usage{
a_line(start, end, ...)
}
\arguments{
\item{start, end}{Start and end of line.}

\item{...}{Any aframe element or named options.}
}
\description{
Add a line.
}
\examples{
browse_aframe(
  a_scene(
    a_dep(),
    a_line("0, 1, 0", "2 0 -5", color = "blue"),
    a_line(" 0 2 0", " 3 0 -4", color = "red")
  )
)

}
