% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{a_scene}
\alias{a_scene}
\title{Scene}
\usage{
a_scene(..., envir = FALSE, anim = FALSE)
}
\arguments{
\item{...}{Any \code{aframer} element or parameter.}

\item{envir}{Set to \code{TRUE} to source \href{https://github.com/feiss/aframe-environment-component/}{aframe-environment-component}.}

\item{anim}{Set to \code{TRUE} to source \href{https://github.com/ngokevin/kframe/tree/master/components/animation/}{aframe-animation-component}.}
}
\description{
Set aframe scene.
}
