% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{a_assets}
\alias{a_assets}
\alias{a_img}
\title{Add assets}
\usage{
a_assets(...)

a_img(id = NULL, src = NULL, ...)
}
\arguments{
\item{...}{Any other parameter.}

\item{id, src}{Id and Source of asset.}
}
\description{
Add assets to your aframe.
}
