% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{attach_aframe}
\alias{attach_aframe}
\title{Attach dependency}
\usage{
attach_aframe(..., envir = FALSE)
}
\arguments{
\item{envir}{Set to \code{TRUE} to source
\href{https://github.com/feiss/aframe-environment-component}{environment components}.}
}
\description{
Attach aframe dependency.
}
