% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{opts_aframe}
\alias{opts_aframe}
\title{Optionise}
\usage{
opts_aframe(...)
}
\arguments{
\item{...}{Any \code{aframer} argument.}
}
\description{
Easily create options.
}
\examples{
(opts_aframe(event = "mouseenter", color = "#8FF7FF"))

}
