% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{a_primitive}
\alias{a_primitive}
\alias{a_entity}
\alias{a_box}
\alias{a_camera}
\alias{a_circle}
\alias{a_collada_model}
\alias{a_cone}
\alias{a_cursor}
\alias{a_curvedimage}
\alias{a_cylinder}
\alias{a_gltf_model}
\alias{a_icosahedron}
\alias{a_image}
\alias{a_light}
\alias{a_link}
\alias{a_obj_model}
\alias{a_octahedron}
\alias{a_plane}
\alias{a_ring}
\alias{a_sky}
\alias{a_sound}
\alias{a_sphere}
\alias{a_tetrahedron}
\alias{a_text}
\alias{a_torus_knot}
\alias{a_torus}
\alias{a_triangle}
\alias{a_video}
\alias{a_videosphere}
\title{Primitives}
\usage{
a_primitive(primitive, opts = list())

a_entity(...)

a_box(...)

a_camera(...)

a_circle(...)

a_collada_model(...)

a_cone(...)

a_cursor(...)

a_curvedimage(...)

a_entity(...)

a_cylinder(...)

a_gltf_model(...)

a_icosahedron(...)

a_image(...)

a_light(...)

a_link(...)

a_obj_model(...)

a_octahedron(...)

a_plane(...)

a_ring(...)

a_ring(...)

a_sky(...)

a_sound(...)

a_sphere(...)

a_tetrahedron(...)

a_text(...)

a_torus_knot(...)

a_torus(...)

a_triangle(...)

a_video(...)

a_videosphere(...)
}
\arguments{
\item{primitive}{The primitive, excluding \code{a-}.}

\item{opts}{A \code{list} of options.}

\item{...}{Any aframe element or named options.}
}
\description{
All aframe primitives.
}
\examples{
x <- a_primitive("box", list(color = "red"))
y <- a_box(color = "red")

identical(x, y)

embed_aframe(
  a_scene(
    a_dep(),
    a_box(
      position = "-1 0.5 -3",
      rotation = "0 45 0",
      color = "#4CC3D9"
    )
  )
)

}
\seealso{
\code{\link{a_scene}} and \code{\link{a_line}}
}
