% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{a_scene}
\alias{a_scene}
\alias{a_dependency}
\title{Scene}
\usage{
a_scene(..., dep = list(version = "0.8.2", cdn = FALSE))

a_dependency(version = "0.8.2", cdn = FALSE)
}
\arguments{
\item{...}{Any \code{aframer} element or parameter.}

\item{dep}{Dependency options, set to \code{NULL} to not import dependency, 
otherwise pass a list containing \code{version} and \code{cdn}.}

\item{version}{Version of aframe to source, see details.}

\item{cdn}{Set to \code{TRUE} to use the CDN and \code{FALSE} to use local file.}
}
\description{
Set aframe scene.
}
\details{
Available versions:
\itemize{
  \item{\code{0.5.0}}
  \item{\code{0.6.0}}
  \item{\code{0.7.0}}
  \item{\code{0.8.0}}
  \item{\code{0.8.2}}
}
}
\note{
It is recommended to import dependency manually, like in the example, rather 
than use the \code{dep} argument.
}
\examples{
scene <- a_scene(
  dep = NULL,
  a_dependency(),
  a_box(
    position = xyz_aframe(0, 1, 0),
    color = "blue"
  )
)

if(interactive())
  browse_aframe(scene)

}
