% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{browse_aframe}
\alias{browse_aframe}
\alias{serve_aframe}
\alias{embed_aframe}
\alias{embed_aframe2}
\title{Browse & Embed}
\usage{
browse_aframe(a)

serve_aframe(a)

embed_aframe(a, width = "100\%", height = "400px")

embed_aframe2(a, width = "100\%", height = "400px")
}
\arguments{
\item{a}{An aframe.}

\item{width, height}{Dimensions of \code{DOM} containing aframe, must be valid \code{CSS}.}
}
\description{
Browse or embed an aframe.
}
\note{
Keep the \code{width} at \code{100\%} for a responsive visualisation.
}
\examples{
browse_aframe(
  a_scene(
    a_box(
      color = "blue",
      position = xyz_aframe(0, 1, -5)
    )
  )
)

}
