% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{a_scene}
\alias{a_scene}
\alias{a_dependency}
\title{Scene}
\usage{
a_scene(..., version = "0.8.2", cdn = FALSE)

a_dependency(version = "0.8.2", cdn = FALSE)
}
\arguments{
\item{...}{Any \code{aframer} element or parameter.}

\item{version}{Version of aframe to source, see details.}

\item{cdn}{Set to \code{TRUE} to use the CDN and \code{FALSE} to use local file.}
}
\description{
Set aframe scene.
}
\details{
Available versions:
\itemize{
  \item{\code{0.5.0}}
  \item{\code{0.6.0}}
  \item{\code{0.7.0}}
  \item{\code{0.8.0}}
  \item{\code{0.8.2}}
}
}
