% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{a_scene}
\alias{a_scene}
\alias{a_dependency}
\title{Scene}
\usage{
a_scene(..., version = "0.8.2", envir = FALSE, anim = FALSE)

a_dependency(version = "0.8.2")
}
\arguments{
\item{...}{Any \code{aframer} element or parameter.}

\item{version}{Version of aframe to source, see details.}

\item{envir}{Set to \code{TRUE} to source \href{https://github.com/feiss/aframe-environment-component/}{aframe-environment-component}.}

\item{anim}{Set to \code{TRUE} to source \href{https://github.com/ngokevin/kframe/tree/master/components/animation/}{aframe-animation-component}.}
}
\description{
Set aframe scene.
}
\details{
Available versions:
\itemize{
  \item{\code{0.7.0}}
  \item{\code{0.8.0}}
  \item{\code{0.8.2}}
}
}
