% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{a_scene}
\alias{a_scene}
\alias{a_dependency}
\alias{a_dep}
\title{Scene}
\usage{
a_scene(...)

a_dependency(version = "0.8.2", cdn = FALSE)

a_dep(version = "0.8.2", cdn = FALSE)
}
\arguments{
\item{...}{Any \code{aframer} element or parameter.}

\item{version}{Version of aframe to source, see details.}

\item{cdn}{Set to \code{TRUE} to use the CDN and \code{FALSE} to use local file.}
}
\description{
Set aframe scene.
}
\details{
Available versions:
\itemize{
  \item{\code{0.5.0}}
  \item{\code{0.6.0}}
  \item{\code{0.7.0}}
  \item{\code{0.8.0}}
  \item{\code{0.8.2}}
}
}
\examples{
scene <- a_scene(
  a_dependency(),
  a_box(
    position = xyz_aframe(0, 1, 0),
    color = "blue"
  )
)

if(interactive())
  browse_aframe(scene)

}
